Add start/end indices to all advection stencils

+++ model/atmosphere/advection/tests/advection_stencil_tests/test_apply_interpolated_tracer_time_tendency.py
@@ -8,6 +8,7 @@
 
 import numpy as np
 import pytest
+from gt4py.next.ffront.fbuiltins import int32
 
 from icon4py.model.atmosphere.advection.stencils.apply_interpolated_tracer_time_tendency import (
     apply_interpolated_tracer_time_tendency,
@@ -45,7 +46,7 @@ class TestApplyInterpolatedTracerTimeTendency(StencilTest):
             p_dtime=p_dtime,
             p_tracer_new=p_tracer_new,
             horizontal_start=0,
-            horizontal_end=grid.num_cells,
+            horizontal_end=int32(grid.num_cells),
             vertical_start=0,
-            vertical_end=grid.num_levels,
+            vertical_end=int32(grid.num_levels),
         )
