+++ model/atmosphere/advection/tests/advection_stencil_tests/test_apply_vertical_density_increment.py
@@ -8,6 +8,7 @@
 
 import numpy as np
 import pytest
+from gt4py.next.ffront.fbuiltins import int32
 
 from icon4py.model.atmosphere.advection.stencils.apply_vertical_density_increment import (
     apply_vertical_density_increment,
@@ -59,7 +60,7 @@ class TestApplyVerticalDensityIncrement(StencilTest):
             p_dtime=p_dtime,
             rhodz_ast2=rhodz_ast2,
             horizontal_start=0,
-            horizontal_end=grid.num_cells,
+            horizontal_end=int32(grid.num_cells),
             vertical_start=0,
-            vertical_end=grid.num_levels,
+            vertical_end=int32(grid.num_levels),
         )
