+++ model/atmosphere/advection/tests/advection_stencil_tests/test_compute_ppm_quartic_face_values.py
@@ -8,6 +8,7 @@
 
 import numpy as np
 import pytest
+from gt4py.next.ffront.fbuiltins import int32
 
 from icon4py.model.atmosphere.advection.stencils.compute_ppm_quartic_face_values import (
     compute_ppm_quartic_face_values,
@@ -71,7 +72,7 @@ class TestComputePpmQuarticFaceValues(StencilTest):
             z_slope=z_slope,
             p_face=p_face,
             horizontal_start=0,
-            horizontal_end=grid.num_cells,
+            horizontal_end=int32(grid.num_cells),
             vertical_start=2,
-            vertical_end=grid.num_levels,
+            vertical_end=int32(grid.num_levels),
         )
